A lot has been done!

+++ src/main/java/com/jamapplicationserver/modules/library/usecases/LibraryEntity/RemoveLibraryEntity/RemoveLibraryEntityUseCase.java
@@ -45,6 +45,10 @@ public class RemoveLibraryEntityUseCase implements IUsecase<RemoveLibraryEntityR
                             .includeUnpublished(request.subjectRole)
                             .getSingleResult();
             if(entity == null) return Result.fail(new LibraryEntityDoesNotExistError());
+            
+            // #### DOMAIN LOGIC LEAK ####
+            if(!request.subjectRole.isAdmin())
+                return Result.fail(new LibraryEntityCannotBeRemovedByAnyoneOtherThanTheAdmin());
 
             repository.remove(entity);
